Fixed logger import paths
