+++ openfl/utilities/utils.py
@@ -65,6 +65,30 @@ def is_api_adress(address: str) -> bool:
     except ValueError:
         return False
 
+def add_log_level(level_name, level_num, method_name=None):
+    """
+    Add a new logging level to the logging module.
+
+    Args:
+        level_name: name of log level.
+        level_num: log level value.
+        method_name: log method wich will use new log level (default = level_name.lower())
+
+    """
+    if not method_name:
+        method_name = level_name.lower()
+
+    def log_for_level(self, message, *args, **kwargs):
+        if self.isEnabledFor(level_num):
+            self._log(level_num, message, args, **kwargs)
+
+    def log_to_root(message, *args, **kwargs):
+        logging.log(level_num, message, *args, **kwargs)
+
+    logging.addLevelName(level_num, level_name)
+    setattr(logging, level_name, level_num)
+    setattr(logging.getLoggerClass(), method_name, log_for_level)
+    setattr(logging, method_name, log_to_root)
 
 
 def validate_file_hash(file_path, expected_hash, chunk_size=8192):
